refactor(v0.3): slim attempt write controller orchestration
